% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-layout.R
\name{plot.pr_layout}
\alias{plot.pr_layout}
\title{Plot a Sensor Layout}
\usage{
\method{plot}{pr_layout}(x, show_regions = TRUE, show_ids = FALSE, ...)
}
\arguments{
\item{x}{A \link{pr_layout} object.}

\item{show_regions}{Logical. If \code{TRUE} (default), draws region masks
in colors. If \code{FALSE}, shows only sensor positions.}

\item{show_ids}{Logical. If \code{TRUE}, labels each sensor with its id.
Default \code{FALSE}.}

\item{...}{Not used.}
}
\value{
A \code{ggplot2} object.
}
\description{
Visualizes a \link{pr_layout} showing active sensor positions and colored
region masks (if any).
}
\examples{
plot(pr_layout_saddle("horse"))
}
