% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-trial.R
\name{plot.pr_trial}
\alias{plot.pr_trial}
\title{Plot a Pressure Trial}
\usage{
\method{plot}{pr_trial}(x, ...)
}
\arguments{
\item{x}{A \link{pr_trial} object.}

\item{...}{Passed to \code{\link[=pr_plot_heatmap]{pr_plot_heatmap()}}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Default plot method for \code{pr_trial} — shows the maximum pressure picture
(MPP) as a heatmap.
}
