% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-frame.R
\name{pr_calc_contact_area}
\alias{pr_calc_contact_area}
\title{Calculate Contact Area Per Frame}
\usage{
pr_calc_contact_area(trial, threshold = 0)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{threshold}{Numeric. Default \code{0}.}
}
\value{
Numeric vector of contact area (cm^2) per frame.
}
\description{
Number of sensors above \code{threshold} times the per-sensor area (cm^2).
}
