% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-frame.R
\name{pr_calc_contact_time}
\alias{pr_calc_contact_time}
\title{Total Contact Time}
\usage{
pr_calc_contact_time(trial, threshold = 0)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{threshold}{Numeric. Default \code{0}.}
}
\value{
Numeric scalar (s).
}
\description{
Duration for which at least one sensor exceeds \code{threshold}.
}
