% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-frame.R
\name{pr_calc_cop_excursion}
\alias{pr_calc_cop_excursion}
\title{COP Excursion (Anterior-Posterior and Medial-Lateral Range)}
\usage{
pr_calc_cop_excursion(trial, threshold = 0)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{threshold}{Numeric. Default \code{0}.}
}
\value{
Named numeric vector with \code{ap} (y-range, mm) and \code{ml} (x-range, mm).
}
\description{
COP Excursion (Anterior-Posterior and Medial-Lateral Range)
}
