% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-frame.R
\name{pr_calc_cop_path}
\alias{pr_calc_cop_path}
\title{Total COP Path Length}
\usage{
pr_calc_cop_path(trial, threshold = 0)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{threshold}{Numeric. Default \code{0}.}
}
\value{
Numeric scalar (mm).
}
\description{
Total COP Path Length
}
