% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-frame.R
\name{pr_calc_force}
\alias{pr_calc_force}
\title{Calculate Total Force Per Frame}
\usage{
pr_calc_force(trial)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}
}
\value{
Numeric vector of length \code{n_frames}.
}
\description{
Force in Newtons equals sum(pressure * sensor_area). With pressure in
kPa and area in cm^2, force[N] = sum(p_kPa * a_cm2) * 0.1.
}
