% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-frame.R
\name{pr_calc_impulse}
\alias{pr_calc_impulse}
\title{Force-Time Integral (Impulse)}
\usage{
pr_calc_impulse(trial)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}
}
\value{
Numeric scalar.
}
\description{
Integrates total force over time. Result is in N*s.
}
