% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-frame.R
\name{pr_calc_loaded_rate}
\alias{pr_calc_loaded_rate}
\title{Calculate Fraction of Loaded Sensors Per Frame}
\usage{
pr_calc_loaded_rate(trial, threshold = 0)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{threshold}{Numeric. Default \code{0}.}
}
\value{
Numeric vector in \verb{[0, 1]} of length \code{n_frames}.
}
\description{
Calculate Fraction of Loaded Sensors Per Frame
}
