% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_downsample}
\alias{pr_downsample}
\title{Downsample a Trial}
\usage{
pr_downsample(trial, factor = 2L)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{factor}{Integer. Downsample factor. Default \code{2}.}
}
\value{
A \link{pr_trial} object with reduced frame count.
}
\description{
Keeps every \code{factor}-th frame.
}
\examples{
tr <- pr_downsample(pr_example_trial("pedar"), factor = 2)
tr$n_frames
}
