% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_filter_time}
\alias{pr_filter_time}
\title{Subset a Trial to a Time Window}
\usage{
pr_filter_time(trial, from = 0, to = Inf)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{from}{Numeric. Start time (s). Default \code{0}.}

\item{to}{Numeric. End time (s). Default \code{Inf}.}
}
\value{
A \link{pr_trial} containing only frames with \code{time} in \verb{[from, to]}.
}
\description{
Subset a Trial to a Time Window
}
\examples{
tr <- pr_filter_time(pr_example_trial("pedar"), from = 1, to = 3)
tr$duration
}
