% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_interpolate}
\alias{pr_interpolate}
\title{Spatial Interpolation of Pressure Data (for Display)}
\usage{
pr_interpolate(trial, factor = 2L)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{factor}{Integer. Interpolation factor in each dimension.
Default \code{2}.}
}
\value{
A list with \code{pressure_interp} (interpolated matrix),
\code{x_mm}, \code{y_mm}.
}
\description{
Performs bilinear interpolation on the maximum pressure picture
between active sensor positions to produce a smoother image. The
underlying trial data is not modified.
}
\examples{
out <- pr_interpolate(pr_example_trial("insole"), factor = 2)
dim(out$pressure_interp)
}
