% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{pr_layout_glove}
\alias{pr_layout_glove}
\title{Get Glove Sensor Layout}
\usage{
pr_layout_glove()
}
\value{
A \link{pr_layout} object.
}
\description{
Returns a \link{pr_layout} for a glove-style hand pressure sensor
(32 sensors across palm and fingertips).
}
