% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{pr_layout_list}
\alias{pr_layout_list}
\title{List All Built-In Sensor Layouts}
\usage{
pr_layout_list()
}
\value{
A \link[tibble:tibble]{tibble::tibble} with columns \code{name}, \code{system}, \code{grid_rows},
\code{grid_cols}, \code{n_sensors}, \code{n_regions}, \code{description}.
}
\description{
List All Built-In Sensor Layouts
}
\examples{
pr_layout_list()
}
