% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{pr_mask_default}
\alias{pr_mask_default}
\title{Get Default Region Masks for a Layout}
\usage{
pr_mask_default(layout)
}
\arguments{
\item{layout}{A \link{pr_layout} object.}
}
\value{
Named list of \link{pr_mask} objects (possibly empty).
}
\description{
Returns a named list of \link{pr_mask} objects built from the \code{regions}
field of a \link{pr_layout}.
}
\examples{
masks <- pr_mask_default(pr_layout_saddle("horse"))
names(masks)
}
