% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{pr_merge_trials}
\alias{pr_merge_trials}
\title{Merge Trials into a Dataset}
\usage{
pr_merge_trials(..., group_var = "condition")
}
\arguments{
\item{...}{\link{pr_trial} objects or lists of them.}

\item{group_var}{Character. Name of a metadata field to use as the
grouping variable.}
}
\value{
A \link{pr_dataset} object.
}
\description{
Merge Trials into a Dataset
}
\examples{
pr_merge_trials(
  pr_example_trial("pedar", seed = 1),
  pr_example_trial("pedar", seed = 2)
)
}
