% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dynamics.R
\name{pr_plot_contact_area}
\alias{pr_plot_contact_area}
\title{Plot Contact Area vs Time}
\usage{
pr_plot_contact_area(trial, show_cycles = FALSE)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{show_cycles}{Logical. If \code{TRUE}, shade detected stance phases.
Default \code{FALSE}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Contact Area vs Time
}
