% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dynamics.R
\name{pr_plot_cop}
\alias{pr_plot_cop}
\title{Plot Center of Pressure Trajectory}
\usage{
pr_plot_cop(trial, show_layout = TRUE, color_by = c("time", "velocity"))
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{show_layout}{Logical. Draw sensor grid in background. Default \code{TRUE}.}

\item{color_by}{Character. \code{"time"} (default) or \code{"velocity"}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Center of Pressure Trajectory
}
\examples{
pr_plot_cop(pr_example_trial("saddle_horse"))
}
