% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dynamics.R
\name{pr_plot_pressure_time}
\alias{pr_plot_pressure_time}
\title{Plot Peak and Mean Pressure vs Time}
\usage{
pr_plot_pressure_time(trial, show_cycles = FALSE)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{show_cycles}{Logical. If \code{TRUE}, shade detected stance phases.
Default \code{FALSE}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Peak and Mean Pressure vs Time
}
\examples{
pr_plot_pressure_time(pr_example_trial("insole"))
}
