% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-regional.R
\name{pr_plot_regional_bar}
\alias{pr_plot_regional_bar}
\title{Regional Parameter Comparison Bar Chart}
\usage{
pr_plot_regional_bar(regional_data, parameter = "mpp", thresholds = NULL)
}
\arguments{
\item{regional_data}{Output of \code{\link[=pr_calc_regional]{pr_calc_regional()}}.}

\item{parameter}{Character. Parameter column to plot. Default \code{"mpp"}.}

\item{thresholds}{A \link[tibble:tibble]{tibble::tibble} from \verb{pr_ref_*()} functions.
If non-\code{NULL}, draws horizontal threshold lines where \code{parameter}
matches.}
}
\value{
A \code{ggplot2} object.
}
\description{
Regional Parameter Comparison Bar Chart
}
\examples{
trial <- pr_example_trial("saddle_horse")
reg <- pr_calc_regional(trial)
pr_plot_regional_bar(reg, "mpp")
}
