% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-regional.R
\name{pr_plot_symmetry}
\alias{pr_plot_symmetry}
\title{Left/Right Symmetry Plot}
\usage{
pr_plot_symmetry(
  trial,
  parameter = c("peak_pressure", "mean_pressure", "force", "contact_area")
)
}
\arguments{
\item{trial}{A \link{pr_trial} object.}

\item{parameter}{Character. One of \code{"peak_pressure"}, \code{"mean_pressure"},
\code{"force"}, \code{"contact_area"}.}
}
\value{
A \code{ggplot2} object with a mirrored bar chart.
}
\description{
Left/Right Symmetry Plot
}
\examples{
pr_plot_symmetry(pr_example_trial("saddle_horse"))
}
