% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{pr_ref_wheelchair}
\alias{pr_ref_wheelchair}
\title{Wheelchair Seating Pressure Thresholds}
\usage{
pr_ref_wheelchair()
}
\value{
A \link[tibble:tibble]{tibble::tibble} with threshold values and sources.
}
\description{
Wheelchair Seating Pressure Thresholds
}
\examples{
pr_ref_wheelchair()
}
