% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-layout.R
\name{pr_validate_layout}
\alias{pr_validate_layout}
\title{Validate a Pressure Sensor Layout}
\usage{
pr_validate_layout(layout)
}
\arguments{
\item{layout}{A \link{pr_layout} object.}
}
\value{
Invisibly returns \code{TRUE} if valid; otherwise throws an informative
error.
}
\description{
Checks internal consistency of a \link{pr_layout} object.
}
\examples{
pr_validate_layout(pr_layout_pedar())
}
